move PKE-on-Azure cluster to CommonCluster adapter to drivers
