+++ internal/providers/azure/pke/driver/commoncluster/cluster.go
